requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.5.7
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.100.0
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
numpy==1.25.1
Pillow==10.0.0
pycountry==22.3.5
pydantic==2.0.3
pydantic_core==2.3.0
pyshp==2.3.1
shapely==2.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1